+++ lib/functions/n_restore_accounts.dart
@@ -1,15 +1,20 @@
 import 'dart:convert';
 
+import 'package:amberflutter/amberflutter.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:ndk/ndk.dart';
+import 'package:ndk_amber/data_layer/data_sources/amber_flutter.dart';
+import 'package:ndk_amber/data_layer/repositories/signers/amber_event_signer.dart';
 import 'package:nip01/nip01.dart';
 import 'package:nip07_event_signer/nip07_event_signer.dart';
+import 'package:nip19/nip19.dart';
 import 'package:nostr_widgets/models/accounts.dart';
 
 Future<void> nRestoreAccounts(Ndk ndk) async {
   final storage = FlutterSecureStorage();
 
   final storedAccounts = await storage.read(key: "nostr_widgets_accounts");
+  print(storedAccounts);
 
   if (storedAccounts == null) return;
 
@@ -26,7 +31,23 @@ Future<void> nRestoreAccounts(Ndk ndk) async {
   }
 
   if (accounts.amber) {
-    // TODO
+    final amber = Amberflutter();
+    final amberFlutterDS = AmberFlutterDS(amber);
+
+    final amberResponse = await amber.getPublicKey();
+
+    final npub = amberResponse['signature'];
+    final pubkey = Nip19.npubToHex(npub);
+
+    final signer = AmberEventSigner(
+      publicKey: pubkey,
+      amberFlutterDS: amberFlutterDS,
+    );
+    ndk.accounts.addAccount(
+      pubkey: pubkey,
+      type: AccountType.externalSigner,
+      signer: signer,
+    );
   }
 
   for (var pubkey in accounts.pubkeys) {