+++ lib/functions/n_save_accounts_state.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:ndk/ndk.dart';
+import 'package:ndk_amber/ndk_amber.dart';
 import 'package:nip07_event_signer/nip07_event_signer.dart';
 import 'package:nostr_widgets/models/accounts.dart';
 
@@ -17,6 +18,11 @@ Future<void> nSaveAccountsState(Ndk ndk) async {
       continue;
     }
 
+    if (account.signer is AmberEventSigner) {
+      accounts.amber = true;
+      continue;
+    }
+
     if (account.type == AccountType.privateKey) {
       final signer = account.signer as Bip340EventSigner;
       if (signer.privateKey == null) continue;
@@ -28,8 +34,6 @@ Future<void> nSaveAccountsState(Ndk ndk) async {
       accounts.pubkeys.add(account.signer.getPublicKey());
       continue;
     }
-
-    // TODO handle amber accounts
   }
 
   accounts.loggedAccount = ndk.accounts.getPublicKey();
@@ -37,6 +41,8 @@ Future<void> nSaveAccountsState(Ndk ndk) async {
   final storage = FlutterSecureStorage();
   await storage.write(
     key: "nostr_widgets_accounts",
-    value: jsonEncode(accounts.toJson()),
+    value: jsonEncode(accounts),
   );
+
+  print(accounts.toJson());
 }