+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 0.0.4
 
+### New features
+
+- Bunker login
+- Nostr connect login
+- Login with Amber
+
 ### Fix
 
 - User picture correctly cover the parent